% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_roxyflow.R
\name{use_roxyflow}
\alias{use_roxyflow}
\title{Use roxyglobals}
\usage{
use_roxyflow()
}
\value{
nothing
}
\description{
Configures roxygen to use \code{\link[=diagram_roclet]{diagram_roclet()}}, adds roxyflow to Suggests
}
\examples{
\dontrun{
use_roxyflow()
}
}
