% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxyflow-package.R
\docType{package}
\name{roxyflow-package}
\alias{roxyflow}
\alias{roxyflow-package}
\title{roxyflow: A `roxygen2` extension for diagramming R packages.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A `roxygen2` extension for diagramming R packages.
}
\author{
\strong{Maintainer}: Patrick Coyle \email{patricktmc@gmail.com}

}
\keyword{internal}
